% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Campinas_daily.R
\docType{data}
\name{Campinas_daily}
\alias{Campinas_daily}
\title{Meteorological data for Campinas, Brazil}
\format{
An object of class \code{data.frame} with 49308 rows and 10 columns.
}
\source{
\url{https://rpubs.com/gabrielblain/1265900}.
}
\usage{
Campinas_daily
}
\description{
Daily meteorological data from Campinas, Sao Paulo State, Brazil.
The meteorological data belongs to the Agronomic Institute of Campinas.
}
\details{
@format ## \code{Campinas_daily}
A data frame with 10 columns and 49308 rows:
\describe{
\item{dates}{date}
\item{year}{years}
\item{month}{months}
\item{days}{days}
\item{Rain}{Rain in millimetres}
\item{tmin}{Minimum air temperature in Celsius degrees}
\item{tmax}{Maximum air temperature in Celsius degrees}
\item{PE}{Potential evapotranspiration in millimetres}
\item{PPE}{Rain minus potential evapotranspiration in millimetres}
\item{julian.day}{julian days}
}
}
\examples{
data(Campinas_daily)
}
\keyword{datasets}
