% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Campinas_monthly.R
\docType{data}
\name{Campinas_monthly}
\alias{Campinas_monthly}
\title{Monthly meteorological data for Campinas, Brazil}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1620 rows and 7 columns.
}
\source{
\url{https://rpubs.com/gabrielblain/1265900}.
}
\usage{
Campinas_monthly
}
\description{
Monthly meteorological data from Campinas, Sao Paulo State, Brazil.
The meteorological data belongs to the Agronomic Institute of Campinas.
}
\details{
@format ## \code{Campinas_monthly}
A data frame with 7 columns and 1620 rows:
\describe{
\item{year}{years}
\item{month}{months}
\item{Rain}{Rain in millimetres}
\item{tmin}{Minimum air temperature in Celsius degrees}
\item{tmax}{Maximum air temperature in Celsius degrees}
\item{PE}{Potential evapotranspiration in millimetres}
\item{PPE}{Rain minus potential evapotranspiration in millimetres}
}
}
\examples{
data(Campinas_monthly)
}
\keyword{datasets}
