% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_PM.R
\name{ET0_PM}
\alias{ET0_PM}
\title{Reference Evapotranspiration Using the Penman and Monteith Method}
\usage{
ET0_PM(Tavg, Tmax, Tmin, Rn, RH, WS, G = NULL, Alt)
}
\arguments{
\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}

\item{Tmax}{A vector, 1-column matrix or data frame with daily maximum air temperature
in Celsius degrees.}

\item{Tmin}{A vector, 1-column matrix or data frame with daily minimum air temperature
in Celsius degrees.}

\item{Rn}{A vector, 1-column matrix or data frame with daily net radiation in
\eqn{MJ m-2 day-1}.}

\item{RH}{A vector, 1-column matrix or data frame with daily relative Humidity in percentage.}

\item{WS}{A vector, 1-column matrix or data frame with daily wind speed in
\eqn{m s-1}.}

\item{G}{Optional. A vector, 1-column matrix or data frame with daily soil heat flux
in \eqn{MJ m-2 day-1}.
Default is \code{NULL} and if \code{NULL} it is assumed to be zero.}

\item{Alt}{A single number defining the altitude in meters.}
}
\value{
A matrix of daily potential evapotranspiration amounts in millimeters.
}
\description{
Calculates daily reference evapotranspiration amounts using the Penman and
Monteith method. This function is imported from the
CropWaterBalance package.
}
\examples{
Tavg <- Campinas[, 2]
Tmax <- Campinas[, 3]
Tmin <- Campinas[, 4]
Rn <- Campinas[, 8]
WS <- Campinas[, 5]
RH <- Campinas[, 6]
ET0_PM(Tavg = Tavg,
       Tmax = Tmax,
       Tmin = Tmin,
       Rn = Rn,
       RH = RH,
       WS = WS,
       Alt = 658)
}
