% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxford.R
\docType{data}
\name{Oxford}
\alias{Oxford}
\title{Daily Air Temperature and Rainfall Data}
\format{
A data frame with 70523 rows and 6 columns:
\describe{
\item{YYYY}{Years}
\item{MM}{Months}
\item{DD}{Days}
\item{Tmax}{Maximum daily air temperature in degrees Celsius}
\item{Tmin}{Minimum daily air temperature in degrees Celsius}
\item{Rainfall}{Precipitation in millimeters}
}
}
\source{
<doi.org/10.6084/m9.figshare.11956239.v1>
}
\usage{
Oxford
}
\description{
Daily meteorological data from Oxford-UK (Radcliffe Observatory).
Covers the period from January 1827 to January 2020.
Belongs to Burt, Stephen (2020). Daily meteorological data from
the Radcliffe Observatory (now Radcliffe Meteorological Station),
from January 1815: updated to January 2020. figshare. Dataset.
}
\keyword{datasets}
