% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPEaggreg_month.R
\name{PPEaggreg_month}
\alias{PPEaggreg_month}
\title{Aggregate monthly precipitation minus potential evapotranspiration (P - PE)
totals at several time scales}
\usage{
PPEaggreg_month(monthly.PPE, start.year, start.month = 1L, TS)
}
\arguments{
\item{monthly.PPE}{Numeric vector, 1-column matrix, or data frame with monthly P - PE totals.}

\item{start.year}{Year at which the aggregation should start.
A single integer.}

\item{start.month}{Month at which the aggregation should start.
A single integer number from 1 to 12. Default is 1, representing January.}

\item{TS}{Integer indicating the time scale on monthly basis
(values between 1 and 24). Default is 1, representing a monthly scale.}
}
\value{
A numeric matrix with columns:
Year, Month, and PPE.at.TS.
The PPE.at.TS column contains the aggregated P - PE values at the specified monthly time scale.
}
\description{
Aggregate monthly precipitation minus potential evapotranspiration (P - PE)
totals at several time scales
}
\examples{
three_month <- PPEaggreg_month(
Campinas_monthly[,7],
start.year = 1890,
start.month = 1,
TS = 3)

}
