% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil_Heat_Flux.R
\name{Soil_Heat_Flux}
\alias{Soil_Heat_Flux}
\title{Soil Heat Flux}
\usage{
Soil_Heat_Flux(Tavg)
}
\arguments{
\item{Tavg}{A vector, 1-column matrix or data frame with daily average air temperature.}
}
\value{
Daily amounts of soil heat flux in \eqn{MJ m-2 day-1}.
}
\description{
Calculates the daily amounts of soil heat flux.
This function is imported from the CropWaterBalance package.
}
\examples{
Tavg <- Campinas[, 2]
Soil_Heat_Flux(Tavg)
}
