% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEIChanges.R
\name{SPEIChanges}
\alias{SPEIChanges}
\title{Detect trends and quantify their effect on the probability of SPEI values occurring}
\usage{
SPEIChanges(PPE.at.TS, nonstat.models = 1, criterion = "BIC")
}
\arguments{
\item{PPE.at.TS}{A 4-column matrix generated as generated by \code{PPESaggreg()}. No other objects are accepted.
\itemize{
\item 1st column is years (YYYY),
\item 2nd is the months (1 to 12),
\item 3rd is the quasiWeeks (1 to 4),
\item and 4th is the difference between P and PE accumulated at a time scale.
}}

\item{nonstat.models}{A single integer number indicating the number of increasingly complex candidate
nonstationary models to be considered in the drought assessments (from 1 to 5).
If 1, only stationary and nonstationary (linear) in location models are fitted (2 candidates).
If 2, a nonstationary (linear) in scale model is included (3 candidates).
If 3, a nonstationary (linear) in both location and scale models is included (4 candidates).
Default is 1.}

\item{criterion}{A character string value (\code{AICc} or \code{BIC}) defining the selection model criterion.
Second-order Akaike Information criterion (\code{AICc}) or Bayesian information criterion (\code{BIC}).
Default is \code{BIC}.}
}
\value{
A \code{list} object with:
\describe{
\item{data.week}{The P - PE amounts, SPEI, cumulative probability of the SPEI values under the stationary
approach, cumulative probability of the SPEI values under the non-stationary approach,
and the changes in the frequency of below zero SPEI values caused by the changes in P - PE patterns.}
\item{Changes.Freq.Drought}{The non-stationary  GEV-based model that best fits the P - PE series,
the expected P - PE amounts describing normal conditions under both stationary and non-stationary
approaches, and the changes in the frequency of moderate to extreme, severe to extreme and extreme drought events}
\item{GEV.parameters}{Parameters of the best fitting GEV model (location, scale and shape) for each quasi-week.}
}
}
\description{
Detect trends and quantify their effect on the probability of SPEI values occurring
}
\examples{
daily.PPE <- Campinas[5480:10958, 9]
PPE.at.TS <- PPEaggreg(daily.PPE, start.date = "1995-01-01", TS = 4)
Changes_SPEI <- SPEIChanges(PPE.at.TS=PPE.at.TS, nonstat.models = 1)
}
