% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEIChanges_month.R
\name{SPEIChanges_month}
\alias{SPEIChanges_month}
\title{Detect trends and quantify their effect on the probability of monthly based SPEI values}
\usage{
SPEIChanges_month(PPE.at.TS, nonstat.models = 1, criterion = "BIC")
}
\arguments{
\item{PPE.at.TS}{A 3-column matrix generated as generated by \code{PPESaggreg_month()}. No other objects are accepted.
\itemize{
\item 1st column is years (YYYY),
\item 2nd is the months (1 to 12),
\item and 3rd is the difference between P and PE accumulated at a monthly based time scale.
}}

\item{nonstat.models}{A single integer number indicating the number of increasingly complex candidate
nonstationary models to be considered in the drought assessments (from 1 to 3).
If 1, only stationary and nonstationary (linear) in location models are fitted (2 candidates).
If 2, a nonstationary (linear) in scale model is included (3 candidates).
If 3, a nonstationary (linear) in both location and scale models is included (4 candidates).
Default is 3.}

\item{criterion}{A character string value (\code{AICc} or \code{BIC}) defining the selection model criterion.
Second-order Akaike Information criterion (\code{AICc}) or Bayesian information criterion (\code{BIC}).
Default is \code{BIC}.}
}
\value{
A \code{list} object with:
\describe{
\item{data.month}{The P - PE amounts, SPEI, cumulative probability of the SPEI values under the stationary
approach, cumulative probability of the SPEI values under the non-stationary approach,
and the changes in the frequency of below zero SPEI values caused by the changes in P - PE patterns.}
\item{Changes.Freq.Drought}{The non-stationary  GEV-based model that best fits the P - PE series,
the expected P - PE amounts describing normal conditions under both stationary and non-stationary
approaches, and the changes in the frequency of moderate to extreme, severe to extreme and extreme drought events}
\item{GEV.parameters}{Parameters of the best fitting GEV model (location, scale and shape) for each quasi-week.}
}
}
\description{
Detect trends and quantify their effect on the probability of monthly based SPEI values
}
\examples{
set.seed(1)
PPE.at.month <- PPEaggreg_month(
  Campinas_monthly[,7],
  start.year = 1890,
  start.month = 1,
  TS = 3
)
Changes_SPEI <- SPEIChanges_month(PPE.at.TS=PPE.at.month,
                                 nonstat.models = 3)

}
