% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_HS.R
\name{ET0_HS}
\alias{ET0_HS}
\title{Potential Evapotranspiration Using Hargreaves-Samani Method}
\usage{
ET0_HS(Ra, Tavg, Tmax, Tmin)
}
\arguments{
\item{Ra}{A \code{vector}, 1-column \code{matrix} or \code{data.frame} with extraterrestrial solar
radiation in \eqn{MJ m-2 day-1}.}

\item{Tavg}{A \code{vector}, 1-column \code{matrix} or \code{data.frame} column with daily average air
temperature.}

\item{Tmax}{A \code{vector}, 1-column \code{matrix} or \code{data.frame} with daily maximum air
temperature in Celsius degrees.}

\item{Tmin}{A \code{vector}, 1-column \code{matrix} or \code{data.frame} with daily minimum air
temperature in Celsius degrees.}
}
\value{
A \code{matrix} of 1-column with the same length as `the input values with the
daily potential evapotranspiration values in millimeters.
}
\description{
Calculates daily reference evapotranspiration amounts using the
Hargreaves-Samani method. This function is imported from the
CropWaterBalance package.
}
\examples{
# See `?Campinas` for more on this data set
Tavg <- Campinas[, 2]
Tmax <- Campinas[, 3]
Tmin <- Campinas[, 4]
Ra <- Campinas[, 7]
ET0_HS(Ra = Ra, Tavg = Tavg, Tmax = Tmax, Tmin = Tmin)
}
