% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Changes_SPEI.R
\name{SPEIChanges}
\alias{SPEIChanges}
\title{Detect trends and quantify their effect on the probability of SPEI values occurring}
\usage{
SPEIChanges(PPE.at.TS, nonstat.models = 1)
}
\arguments{
\item{PPE.at.TS}{A 4-column matrix generated with \code{PPESaggreg()}. No other objects are accepted.
\itemize{
\item 1st column is years (YYYY),
\item 2nd is the months (1 to 12),
\item 3rd is the quasiWeeks (1 to 4),
\item and 4th is the difference between P and PE accumulated at a time scale.
}}

\item{nonstat.models}{A single integer number indicating the number of increasingly complex candidate
nonstationary models to be considered in the drought assessments (from 1 to 5).
If 1, only stationary and nonstationary (linear) in location models are fitted (2 candidates).
If 2, a nonstationary (linear) in scale model is included (3 candidates).
If 3, a nonstationary (linear) in both location and scale models is included (4 candidates).
If 4, a nonstationary (polynomial 2nd degree) in location model is included (5 candidates).
If 5, a nonstationary (polynomial 2nd degree) in location and linear in scale model is included (6 candidates).
Default is 1.}
}
\value{
A \code{list} object with:
\describe{
\item{data.week}{The P - PE amounts, SPEI, cumulative probability of the SPEI values under the stationary
approach, cumulative probability of the SPEI values under the non-stationary approach,
and the changes in the frequency of below zero SPEI values caused by the changes in P - PE patterns.}
\item{model.selection}{The non-stationary  GEV-based model that best fits the P - PE series}
\item{Changes.Freq.Drought}{changes in the frequency of moderate to extreme,
severe to extreme and extreme drought events,as categorized by the SPEI classification system,
caused by the changes in P - PE patterns.
Changes in the climate expected P - PE amount are also shown.}
\item{Statistics}{Year to year changes in the expected frequency of moderate to extreme, severe to extreme
and extreme drought events.}
\item{data.week}{P - PE amounts, \acronym{SPEI}, cumulative
probability of the \acronym{SPEI} values under the stationary approach,
cumulative probability of the \acronym{SPEI} values under the non-stationary
approach, and the changes in the frequency of below zero \acronym{SPEI}
values caused by the changes in P - PE patterns.}
\item{model.selection}{The generalized additive model that best fits the
P - PE series}
\item{Changes.Freq.Drought}{changes in the frequency of zero P - PE,
moderate, severe and extreme drought events, as defined by the
\acronym{SPEI} classification system, caused by the changes in P - PE
patterns. Changes in the P - PE amounts associated describing normal
conditions is also shown.}
\item{Statistics}{Year to year changes in the expected frequency of
moderate, severe and extreme drought events.}
}
}
\description{
Detect trends and quantify their effect on the probability of SPEI values occurring
}
\details{
The \code{SPEIChanges()} function implements a nonstationary parametric approach to detect
changes in balance between precipitation and potential evapotranspitation patterns and
assess their impact on the expected frequency of Standardized Precipitation Evapotranspiration
Index values (SPEI). It evaluates up tp 6 candidate models
based on time-varying GEV distributions, which account for a broad range of linear
and nonlinear changes in both the mean and dispersion of the P - PE series.

Model selection is performed using the second-order Akaike Information Criterion (AICc),
and the selected model is used to compute the cumulative probability of each
P - PE amount under changing climate conditions. These nonstationary probabilities
are then compared with those from the original, stationary SPEI algorithm to identify
whether the frequency of drought events has increased or decreased over time.

For detailed explanations of the gamma-based models and the model selection procedure,
please refer to the README and Vignettes included in the package.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

PPETS4 <- PPETS4
Changes_SPEI <- SPEIChanges(PPE.at.TS=PPETS4, nonstat.models = 1)
\dontshow{\}) # examplesIf}
}
